pre-launch production ready check
= 1.0.0 =
* Adjusted padding and h2 margin for site info block
* no-label class for ACF
* Changed li bullet styles
* Adjusted H font sizes
* Added author filter

+++ templates/content-page.php
@@ -1,15 +1,16 @@
 <section>
 <?php
-
-if (get_field('video')) {
-  echo '<div class="row">';
-  echo '<div class="col">';
-  the_content();
-  echo '</div>';
-  echo '<div class="col">';
-  the_field('video');
-  echo '</div>';
-  echo '</div>';
+if (class_exists('acf')) {
+  if (get_field('video')) {
+    echo '<div class="row">';
+    echo '<div class="col">';
+    the_content();
+    echo '</div>';
+    echo '<div class="col">';
+    the_field('video');
+    echo '</div>';
+    echo '</div>';
+  }
 } else {
   the_content();
 }
